working login+openai and bedrock
