+++ server/controllers/dataController.ts
@@ -27,29 +27,5 @@ const dataController = {
       });
     }
   },
-  async getAllUserWebsites(req: Request, res: Response, next: NextFunction) {
-    try {
-      const id = res.locals.userId;
-
-      const response = await pool.query(
-        `SELECT DISTINCT website_name
-         FROM "clickTable2"
-         WHERE user_id = $1`,
-        [id]
-      );
-
-      if (response.rows.length > 0) {
-        res.status(200).json(response.rows);
-      } else {
-        res.status(404).json({ message: "No data" });
-      }
-    } catch (err) {
-      const error = err as Error;
-      return next({
-        message: "Error in getAllUserWebsites: " + error.message,
-        log: err,
-      });
-    }
-  },
 };
 export default dataController;
