+++ server/models/db.ts
@@ -1,6 +1,6 @@
 const { Pool } = require('pg');
 //This should be put in an .env, right? - Eric 08/23/24
-const URI = 'postgresql://postgres.ltmylkgzewtbyzppqgzn:ActivityTracker123!@aws-0-us-west-1.pooler.supabase.com:6543/postgres'
+const URI = process.env.DB_URI_STRING
 
 const pool = new Pool({
     connectionString: URI,
