+++ server/routes/dataRoute.ts
@@ -5,6 +5,5 @@ import authMiddleware from "../middleware/auth";
 
 
 router.get("/",authMiddleware, dataController.getAllUserData,(req: Request, res: Response) => {});
-router.get("/websites",authMiddleware, dataController.getAllUserWebsites,(req: Request, res: Response) => {});  
 
 export default router;
